TR-5032 Refactor run operation reponses (#20)
Change the return type of Transosit.run to simplify it.

Return value now contains just the requestId and the array of values
returned. It also includes a convenience getter for a single value if
the called operation should only return a value. (Even if an operation
returns a non-array, we wrap the result in an array in the response.)

Non-successful operations throw an exception, so result only needs to
handle successful cases. Code that calls operations that might fail
should catch and handle these cases.

Add type parameterization to run function, so caller can declare the
expected form of the JSON response. In the future, we may limit type
parameterization to only allow values that one could represent as JSON,
but as of now we don't know how to do that correctly in TypeScript's
types-ystem.